requirements.txt
discord.py[voice]
discord-ext-voice-recv
python-dotenv
ffmpeg-python
requests
groq
PyNaCl
SpeechRecognition
audioop-lts
gTTS
pyttsx3
